backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Firebase
firebase-admin==6.4.0

# LiveKit
livekit==0.12.0
livekit-agents==0.8.3

# Speech & LLM
deepgram-sdk==3.2.5
groq==0.4.2
edge-tts==6.1.9

# PDF & NLP
pdfplumber==0.10.3
PyPDF2==3.0.1
spacy==3.7.2
scikit-learn==1.3.2
sentence-transformers==2.2.2

# Report Generation
reportlab==4.0.8
matplotlib==3.8.2

# Utilities
httpx==0.26.0
aiofiles==23.2.1
websockets==12.0